+++ src/pages/OrderPage/orders-list/OrdersList.jsx
@@ -1,18 +1,38 @@
+import { useEffect } from "react";
+import { useAppDispatch, useAppSelector } from "../../../hooks/redux";
+import { useAuth } from "../../../hooks/useAuth";
 import styles from "./OrdersList.module.scss";
 import OrderItem from "./order-item/OrderItem";
+import { fetchOrder } from "../../../store/order/order.slice";
+import CartEmpty from "../../../components/cartEmpty/CartEmpty";
 
 const OrdersList = () => {
+  const { id } = useAuth();
+  const { order } = useAppSelector((state) => state.orderSlice);
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    dispatch(fetchOrder(id));
+  }, [id]);
+
+  if (!order.length) {
+    return <CartEmpty title="주문내역" />;
+  }
   return (
     <div className={styles.orders}>
-      <div>
-        <div className={styles.order_header}>
-          <h3>주문 번호</h3>
-          <p>합계: $</p>
+      {order.map((item) => (
+        <div key={item.id}>
+          <div className={styles.order_header}>
+            <h3>주문 번호_{item.id}</h3>
+            <p>합계: $ {item.totalPrice.toFixed(2)}</p>
+          </div>
+          <ul className={styles.orders_list}>
+            {item.products.map((order) => (
+              <OrderItem key={order.id} order={order} />
+            ))}
+          </ul>
         </div>
-        <ul className={styles.orders_list}>
-          <OrderItem />
-        </ul>
-      </div>
+      ))}
     </div>
   );
 };
