+++ src/pages/OrderPage/orders-list/order-item/OrderItem.jsx
@@ -1,21 +1,25 @@
 import styles from "./OrderItem.module.scss";
+import { Link } from "react-router-dom";
 
-const OrderItem = () => {
+const OrderItem = ({ order }) => {
   return (
     <li className={styles.order_item}>
-      <img src="주문이미지" alt="product card" />
-
+      <Link to={`/product/${order.id}`}>
+        <img src={order.image} alt="product card" />
+      </Link>
       <div className={styles.order_description}>
-        <h4>주문카테고리</h4>
-        <h3>주문타이틀</h3>
+        <h4>{order.category}</h4>
+        <h3>{order.title}</h3>
       </div>
       <div className={styles.order_price}>
         <h4>가격 :</h4>
-        <span>$ 주문가격 x 주문갯수 </span>
+        <span>
+          $ {order.price} x {order.quantity}{" "}
+        </span>
       </div>
       <div className={styles.order_total}>
         <h4>합계 : </h4>
-        <span>$ 주문합계</span>
+        <span>$ {order.total}</span>
       </div>
     </li>
   );
